[UIK-5559][chore] fix NVDA in GH Actions
